Split profiles into separate modules
