+++ structuralcodes/geometry/profiles/__init__.py
@@ -1,6 +1,12 @@
 """Main entry point for profiles."""
 
-from ._steel_sections import HE, IPE, IPN, UB, UBP, UC, UPN
+from ._he import HE
+from ._ipe import IPE
+from ._ipn import IPN
+from ._ub import UB
+from ._ubp import UBP
+from ._uc import UC
+from ._upn import UPN
 
 __all__ = [
     'HE',
